# 灵月仙途/api-definitions.yaml
openapi: 3.0.0
info:
  title: 灵月仙途 API
  description: 灵月仙途游戏后端API接口定义
  version: 1.0.0
servers:
  - url: http://localhost:8088/api
    description: 本地开发环境
paths:
...
  /role/{roleId}:
    get:
      summary: 获取角色详情
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameRole'

  # 背包相关API
...
  /clan/member/role/{roleId}:
    get:
      summary: 获取角色的宗门成员信息
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClanMember'
...
  /role-skill/role/{roleId}:
    get:
      summary: 获取角色技能
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleSkill'
components:
  schemas:
...
    GameRole:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        roleName:
          type: string
        gender:
          type: integer
        realm:
          type: string
        realmStage:
          type: string
        worldLevel:
          type: string
        cultivation:
          type: integer
        cultivationMax:
          type: integer
        spiritStones:
          type: integer
        hp:
          type: integer
        mp:
          type: integer
        spiritRoot:
          type: string
        profession:
          type: string
        professionLevel:
          type: integer
        professionExp:
          type: integer
        professionCertified:
          type: boolean
        sect:
          type: string
        sectPosition:
          type: string
        partner:
          type: string
        todayCultivationTimes:
          type: integer
        maxDailyCultivationTimes:
          type: integer
        createTime:
          type: string
          format: date-time

    # 背包相关
...
    ClanMember:
      type: object
      properties:
        id:
          type: integer
          format: int64
        clanId:
          type: integer
          format: int64
        roleId:
          type: integer
          format: int64
        roleName:
          type: string
        position:
          type: string
        joinTime:
          type: string
          format: date-time
        contribution:
          type: integer
...
    RoleSkill:
      type: object
      properties:
        id:
          type: integer
          format: int64
        roleId:
          type: integer
          format: int64
        skillName:
          type: string
        skillType:
          type: string
        level:
          type: integer
        exp:
          type: integer
        maxExp:
          type: integer
        effect:
          type: string
        unlocked:
          type: boolean